requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
gunicorn==25.1.0
google-api-python-client
google-auth
google-auth-httplib2
psycopg2-binary
gspread
google-auth
